Add code for HW1

+++ LinearEq.java
@@ -9,7 +9,7 @@ public class LinearEq {
 		double a= Double.parseDouble(args[0]);
 		double b= Double.parseDouble(args[1]);
 		double c= Double.parseDouble(args[2]);
-		double solution= (c-b)/a;
+		double solution= (double)(c-b)/a;
 		System.out.println(a+" * x + "+b+" = "+c);
 		System.out.println("X = "+solution);
